+++ src/plotting.jl
@@ -15,27 +15,26 @@ function heatmap_basins_attractors end
 function heatmap_basins_attractors! end
 export heatmap_basins_attractors, heatmap_basins_attractors!
 
-""" 
-    shaded_basins_heatmap(grid, basins, iterations, attractors; kwargs...) 
+"""
+    shaded_basins_heatmap(grid, basins, attractors, iterations; kwargs...)
 
 Plot a heatmap of found (2-dimensional) `basins` of attraction and corresponding
 `attractors`. An matrix `iterations` with the same size of `basins` must be provided
-to shade the color according to the value of this matrix. An small value corresponds 
+to shade the color according to the value of this matrix. An small value corresponds
 to a light color and a large value to a darker tone. This is useful to represent
-the number of iterations taken for each initial conditions. See also 
-[`get_iterations`]@ref to store this iteration number.
+the number of iterations taken for each initial conditions. See also
+[`iterations_to_converge`]@ref to store this iteration number.
 
 ## Keyword arguments
 - `show_attractors = true`: shows the attractor on plot
-- `maxit = maximum(iterations)`: clip the values of the matrix iterations to 
-the value `maxit`. Useful when there are some very long iterations and keep the 
-range constrained to a give interval.
+- `maxit = maximum(iterations)`: clip the values of `iterations` to
+the value `maxit`. Useful when there are some very long iterations and keep the
+range constrained to a given interval.
 - All the [common plotting keywords](@ref).
 """
 function shaded_basins_heatmap end
 function shaded_basins_heatmap! end
 export shaded_basins_heatmap, shaded_basins_heatmap!
-   
 
 ##########################################################################################
 # Continuation